requirements.txt
discord.py
pymongo
dnspython
dblpy